+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 ### Added
 
 - Link to the front page of the current locale to the logo in the footer.
+- Link to the online store of the company to the footer.
 
 ### Changed